+++ apps/mobile/src/components/NostrVideo/styles.ts
@@ -20,7 +20,7 @@ export default ThemedStyleSheet(() => ({
   actionsContainer: {
     position: 'relative',
     right: Dimensions.get('window').width * 0.1,
-    bottom: Dimensions.get('window').height * 0.1,
+    bottom: Dimensions.get('window').height * 0.15,
     // right: Dimensions.get('window').width * 0.2,
     // bottom: Dimensions.get('window').height * 0.2,
     display: 'flex',
