+++ apps/nestjs-indexer/src/common/constants.ts
@@ -18,8 +18,9 @@ export default {
       //   '0x4a5d414ec6085ef5d3779b6e72015076b1a1ecdb6065fe1123d794ff41aea4d',
       // LAUNCHPAD_ADDRESS:
       //   '0x2b4d93fc565381d1911f3f449f615de050b72d297fc95d89dda0301d7d35a37',
-      LAUNCHPAD_ADDRESS:
-        '0x7fbf067657772a454c302354a19e07ce0a920736e2e3b7ca605d813723db883',
+      // LAUNCHPAD_ADDRESS:
+      // '0x7fbf067657772a454c302354a19e07ce0a920736e2e3b7ca605d813723db883',
+      LAUNCHPAD_ADDRESS:"0x45cb52efdb2f80811c461f7115e550574739a6cd2725df759022899286b9dc1",
       NAMESERVICE_ADDRESS:
         '0x15dcd3c28c07846fa98d3a40d29446de21b5e6cd8d49a43773da0f237d5ea7f',
       ESCROW_DEPOSIT_ADDRESS:
