+++ packages/afk_nostr_sdk/src/hooks/videos/useGetVideos.ts
@@ -19,16 +19,25 @@ export const useGetVideos = (options?: UseRootNotesOptions) => {
     getNextPageParam: (lastPage: any, allPages, lastPageParam) => {
       if (!lastPage?.length) return undefined;
 
-      const pageParam = lastPage[lastPage.length - 1].created_at - 1;
+      const lastNote = lastPage[lastPage.length - 1];
+      // console.log('created_at', lastNote.created_at);
+      const pageParam = lastNote.created_at - 1;
+      // console.log('pageParam', pageParam);
 
-      if (!pageParam || pageParam === lastPageParam) return undefined;
       return pageParam;
+      // if (!lastPage?.length) return undefined;
+
+      // const pageParam = lastPage[lastPage.length - 1].created_at - 1;
+
+      // if (!pageParam || pageParam === lastPageParam) return undefined;
+
+      // return pageParam;
     },
     queryFn: async ({ pageParam }) => {
 
       const sinceTimestamp = pageParam
-        ? pageParam - 1 * 60 * 60 :// Restart from pageParam minus 1 hour
-        Math.round(Date.now() / 1000) - 1 * 60 * 60; // Start from 1 hour ago
+        ? pageParam - 1 * 60 * 60 * 24 :// Restart from pageParam minus 1 hour
+        Math.round(Date.now() / 1000) - 1 * 60 * 60 * 24; // Start from 1 hour ago
 
 
       const notes = await ndk.fetchEvents({
