+++ scripts/.env.exemple
@@ -39,7 +39,8 @@ ART_PEACE_CLASS_HASH=0x05a1a0a58b486db1400c1c443ac9360ff0e83d96be2ce80af33e59634
 # LAUNCHPAD_CLASS_HASH=0x01d2173db2f78de9d202be552d66684e3e13553a237af35d826434fab4920edc
 # CREATOR FEE ENABLED
 # LAUNCHPAD_CLASS_HASH=0x61292c14547c80c0f7c3e1af0d74c400db5e363cc8be5fc8e8c9fd031b4bc57
-LAUNCHPAD_CLASS_HASH=0x13df1221d7e05d793bb0c09e651010a47b5961057d1ac3673fe18147ad104f6
+# LAUNCHPAD_CLASS_HASH=0x13df1221d7e05d793bb0c09e651010a47b5961057d1ac3673fe18147ad104f6
+LAUNCHPAD_CLASS_HASH=0x78dd6dc1bc0961ccbfa9cb939ff2e5d35ac31f680ca831c3bf420ba4145c2db
 
 USERNAME_STORE_CLASS_HASH=0x30b77a9e4d3a2fac92cc23b609416fef29926cfa36d01f969a933f95ee5ad16
 MEMECOIN_CLASS_HASH=0x43b100e1cd969dc45b1d59afcda47ee52444e5baba9ef511a16858d2067ecb7
@@ -48,7 +49,8 @@ TOKEN_CLASS_HASH=0x97341df71bdc18c3c1d9496a238f92b895c0c3d61725481641c851d3db085
 CANVAS_NFT_CLASS_HASH=0x30b77a9e4d3a2fac92cc23b609416fef29926cfa36d01f969a933f95ee5ad16
 BASE_URI="The pixel backend"
 
-UNRUG_LIQUIDITY_CLASS_HASH=0x3847027c61b109ede54a3d5ac6251e409d425984536204ce7a908104fdc0904
+# UNRUG_LIQUIDITY_CLASS_HASH=0x3847027c61b109ede54a3d5ac6251e409d425984536204ce7a908104fdc0904
+UNRUG_LIQUIDITY_CLASS_HASH=0x379b557be4366c38a674dd7193c693ccb4ab2c6323e8325d3625c8efcc2e0c3
 
 
 # DAO
